Platform class,
Added a new folder for the platform class, Added array for platforms which adds a 100px space inbetween each platforms (which does not work yet).  Draw() in the platform class which is for moving platforms, if it hits the wall at x > 500 it will start to move in the oppisite direction and vice versa for x + width.  Trying to draw the platforms in the draw screen right now. Ball also stopped moving